requirements.txt
streamlit
google-generativeai
python-dotenv
supabase
BeautifulSoup4
lxml
faust-cchardet
bcrypt
json-repair
PyJWT
streamlit_extras
extra-streamlit-components
pandas